refactor: create a view module and split each view into a file
